// restore/lib/types.ts
export interface Maquininha {
  id: string;
  nome: string;
  taxa_debito: number;
  taxa_credito: number;
  taxa_credito_parcelado: number;
  ativo: boolean;
  created_at: string;
}